remove legacy blocks and refactor within the module itself.
