--- src/diffusers/models/transformers/legacy/__init__.py
@@ -1,3 +1,0 @@
-from .transformer_2d_continuous import ContinuousTransformer2DModel
-from .transformers_2d_patched import PatchedTransformer2DModel
-from .transformers_2d_vectorized import VectorizedTransformer2DModel
