+++ src/diffusers/models/transformers/transformer_2d.py
@@ -12,15 +12,18 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 from dataclasses import dataclass
-from typing import Optional
+from typing import Any, Dict, Optional
 
 import torch
-import torch.nn as nn
+import torch.nn.functional as F
+from torch import nn
 
 from ...configuration_utils import ConfigMixin, register_to_config
-from ...utils import BaseOutput, deprecate, logging
+from ...utils import BaseOutput, deprecate, is_torch_version, logging
 from ..attention import BasicTransformerBlock
+from ..embeddings import ImagePositionalEmbeddings, PatchEmbed, PixArtAlphaTextProjection
 from ..modeling_utils import ModelMixin
+from ..normalization import AdaLayerNormSingle
 
 
 logger = logging.get_logger(__name__)  # pylint: disable=invalid-name
@@ -147,23 +150,120 @@ class Transformer2DModel(ModelMixin, ConfigMixin):
                 f" {patch_size}. Make sure that `in_channels`, `num_vector_embeds` or `num_patches` is not None."
             )
 
-        out_channels = in_channels if out_channels is None else out_channels
+        self.in_channels = in_channels
 
-        # Initialize the appropriate variant and return.
+        # 2. Initialize the right blocks.
         if self.is_input_continuous:
-            input_type = "continuous"
+            self._init_continuous_input(
+                in_channels,
+                out_channels,
+                inner_dim,
+                num_attention_heads,
+                attention_head_dim,
+                cross_attention_dim,
+                activation_fn,
+                num_embeds_ada_norm,
+                attention_bias,
+                only_cross_attention,
+                double_self_attention,
+                upcast_attention,
+                norm_type,
+                norm_elementwise_affine,
+                norm_eps,
+                attention_type,
+                norm_num_groups,
+                use_linear_projection,
+                dropout,
+                num_layers,
+            )
+        elif self.is_input_vectorized:
+            self._init_vectorized_inputs(
+                in_channels,
+                sample_size,
+                num_vector_embeds,
+                inner_dim,
+                num_attention_heads,
+                attention_head_dim,
+                dropout,
+                cross_attention_dim,
+                activation_fn,
+                num_embeds_ada_norm,
+                attention_bias,
+                only_cross_attention,
+                double_self_attention,
+                upcast_attention,
+                norm_type,
+                norm_elementwise_affine,
+                norm_eps,
+                attention_type,
+                num_layers,
+            )
         elif self.is_input_patches:
-            input_type = "patched"
-        elif self.is_input_vectorized == "vectorized":
-            input_type = "vectorized"
+            self._init_patched_inputs(
+                in_channels,
+                out_channels,
+                sample_size,
+                patch_size,
+                inner_dim,
+                num_attention_heads,
+                attention_head_dim,
+                dropout,
+                cross_attention_dim,
+                activation_fn,
+                num_embeds_ada_norm,
+                attention_bias,
+                only_cross_attention,
+                double_self_attention,
+                upcast_attention,
+                norm_type,
+                norm_elementwise_affine,
+                norm_eps,
+                attention_type,
+                num_layers,
+                interpolation_scale,
+                caption_channels,
+            )
 
-        return Transformer2DModel._create_model(
-            input_type=input_type,
-            in_channels=in_channels,
-            out_channels=out_channels,
+        # Others.
+        self.out_channels = in_channels if out_channels is None else out_channels
+        self.gradient_checkpointing = False
+
+    def _init_continuous_input(
+        self,
+        in_channels,
+        out_channels,
+        inner_dim,
+        num_attention_heads,
+        attention_head_dim,
+        cross_attention_dim,
+        activation_fn,
+        num_embeds_ada_norm,
+        attention_bias,
+        only_cross_attention,
+        double_self_attention,
+        upcast_attention,
+        norm_type,
+        norm_elementwise_affine,
+        norm_eps,
+        attention_type,
+        norm_num_groups,
+        use_linear_projection,
+        dropout,
+        num_layers,
+    ):
+        self.use_linear_projection = use_linear_projection
+
+        self.norm = torch.nn.GroupNorm(num_groups=norm_num_groups, num_channels=in_channels, eps=1e-6, affine=True)
+        if use_linear_projection:
+            self.proj_in = torch.nn.Linear(in_channels, inner_dim)
+        else:
+            self.proj_in = torch.nn.Conv2d(in_channels, inner_dim, kernel_size=1, stride=1, padding=0)
+
+        self.transformer_blocks = self._get_transformer_blocks(
             inner_dim=inner_dim,
             num_attention_heads=num_attention_heads,
             attention_head_dim=attention_head_dim,
+            dropout=dropout,
             cross_attention_dim=cross_attention_dim,
             activation_fn=activation_fn,
             num_embeds_ada_norm=num_embeds_ada_norm,
@@ -175,68 +275,396 @@ class Transformer2DModel(ModelMixin, ConfigMixin):
             norm_elementwise_affine=norm_elementwise_affine,
             norm_eps=norm_eps,
             attention_type=attention_type,
-            norm_num_groups=norm_num_groups,
-            use_linear_projection=use_linear_projection,
+            num_layers=num_layers,
+        )
+        # TODO: should use out_channels for continuous projections
+        if use_linear_projection:
+            self.proj_out = torch.nn.Linear(inner_dim, in_channels)
+        else:
+            self.proj_out = torch.nn.Conv2d(inner_dim, in_channels, kernel_size=1, stride=1, padding=0)
+
+    def _init_vectorized_inputs(
+        self,
+        in_channels,
+        sample_size,
+        num_vector_embeds,
+        inner_dim,
+        num_attention_heads,
+        attention_head_dim,
+        dropout,
+        cross_attention_dim,
+        activation_fn,
+        num_embeds_ada_norm,
+        attention_bias,
+        only_cross_attention,
+        double_self_attention,
+        upcast_attention,
+        norm_type,
+        norm_elementwise_affine,
+        norm_eps,
+        attention_type,
+        num_layers,
+    ):
+        assert sample_size is not None, "Transformer2DModel over discrete input must provide sample_size"
+        assert num_vector_embeds is not None, "Transformer2DModel over discrete input must provide num_embed"
+
+        self.height = sample_size
+        self.width = sample_size
+        self.num_vector_embeds = num_vector_embeds
+        self.num_latent_pixels = self.height * self.width
+
+        self.latent_image_embedding = ImagePositionalEmbeddings(
+            num_embed=num_vector_embeds, embed_dim=inner_dim, height=self.height, width=self.width
+        )
+
+        self.transformer_blocks = self._get_transformer_blocks(
+            inner_dim=inner_dim,
+            num_attention_heads=num_attention_heads,
+            attention_head_dim=attention_head_dim,
             dropout=dropout,
+            cross_attention_dim=cross_attention_dim,
+            activation_fn=activation_fn,
+            num_embeds_ada_norm=num_embeds_ada_norm,
+            attention_bias=attention_bias,
+            only_cross_attention=only_cross_attention,
+            double_self_attention=double_self_attention,
+            upcast_attention=upcast_attention,
+            norm_type=norm_type,
+            norm_elementwise_affine=norm_elementwise_affine,
+            norm_eps=norm_eps,
+            attention_type=attention_type,
             num_layers=num_layers,
         )
 
-    @classmethod
-    def _create_model(cls, input_type, **kwargs):
-        if input_type == "continuous":
-            from .legacy import ContinuousTransformer2DModel
+        self.norm_out = nn.LayerNorm(inner_dim)
+        self.out = nn.Linear(inner_dim, self.num_vector_embeds - 1)
 
-            return ContinuousTransformer2DModel(**kwargs)
-        elif input_type == "vectorized":
-            from .legacy import VectorizedTransformer2DModel
+    def _init_patched_inputs(
+        self,
+        in_channels,
+        out_channels,
+        sample_size,
+        patch_size,
+        inner_dim,
+        num_attention_heads,
+        attention_head_dim,
+        dropout,
+        cross_attention_dim,
+        activation_fn,
+        num_embeds_ada_norm,
+        attention_bias,
+        only_cross_attention,
+        double_self_attention,
+        upcast_attention,
+        norm_type,
+        norm_elementwise_affine,
+        norm_eps,
+        attention_type,
+        num_layers,
+        interpolation_scale,
+        caption_channels,
+    ):
+        assert sample_size is not None, "Transformer2DModel over patched input must provide sample_size"
 
-            return VectorizedTransformer2DModel(**kwargs)
-        elif input_type == "patched":
-            from .legacy import PatchedTransformer2DModel
+        self.height = sample_size
+        self.width = sample_size
 
-            return PatchedTransformer2DModel(**kwargs)
-        else:
-            raise ValueError(f"Invalid input_type: {input_type}")
+        self.patch_size = patch_size
+        interpolation_scale = (
+            interpolation_scale if interpolation_scale is not None else max(self.config.sample_size // 64, 1)
+        )
+        self.pos_embed = PatchEmbed(
+            height=sample_size,
+            width=sample_size,
+            patch_size=patch_size,
+            in_channels=in_channels,
+            embed_dim=inner_dim,
+            interpolation_scale=interpolation_scale,
+        )
 
+        self.transformer_blocks = self._get_transformer_blocks(
+            inner_dim=inner_dim,
+            num_attention_heads=num_attention_heads,
+            attention_head_dim=attention_head_dim,
+            dropout=dropout,
+            cross_attention_dim=cross_attention_dim,
+            activation_fn=activation_fn,
+            num_embeds_ada_norm=num_embeds_ada_norm,
+            attention_bias=attention_bias,
+            only_cross_attention=only_cross_attention,
+            double_self_attention=double_self_attention,
+            upcast_attention=upcast_attention,
+            norm_type=norm_type,
+            norm_elementwise_affine=norm_elementwise_affine,
+            norm_eps=norm_eps,
+            attention_type=attention_type,
+            num_layers=num_layers,
+        )
 
-def _get_transformer_blocks(
-    inner_dim,
-    num_attention_heads,
-    attention_head_dim,
-    dropout,
-    cross_attention_dim,
-    activation_fn,
-    num_embeds_ada_norm,
-    attention_bias,
-    only_cross_attention,
-    double_self_attention,
-    upcast_attention,
-    norm_type,
-    norm_elementwise_affine,
-    norm_eps,
-    attention_type,
-    num_layers,
-):
-    transformer_blocks = nn.ModuleList(
-        [
-            BasicTransformerBlock(
-                inner_dim,
-                num_attention_heads,
-                attention_head_dim,
-                dropout=dropout,
-                cross_attention_dim=cross_attention_dim,
-                activation_fn=activation_fn,
-                num_embeds_ada_norm=num_embeds_ada_norm,
-                attention_bias=attention_bias,
-                only_cross_attention=only_cross_attention,
-                double_self_attention=double_self_attention,
-                upcast_attention=upcast_attention,
-                norm_type=norm_type,
-                norm_elementwise_affine=norm_elementwise_affine,
-                norm_eps=norm_eps,
-                attention_type=attention_type,
+        if norm_type != "ada_norm_single":
+            self.norm_out = nn.LayerNorm(inner_dim, elementwise_affine=False, eps=1e-6)
+            self.proj_out_1 = nn.Linear(inner_dim, 2 * inner_dim)
+            self.proj_out_2 = nn.Linear(inner_dim, patch_size * patch_size * self.out_channels)
+        elif norm_type == "ada_norm_single":
+            self.norm_out = nn.LayerNorm(inner_dim, elementwise_affine=False, eps=1e-6)
+            self.scale_shift_table = nn.Parameter(torch.randn(2, inner_dim) / inner_dim**0.5)
+            self.proj_out = nn.Linear(inner_dim, patch_size * patch_size * self.out_channels)
+
+        # PixArt-Alpha blocks.
+        self.adaln_single = None
+        self.use_additional_conditions = False
+        if norm_type == "ada_norm_single":
+            self.use_additional_conditions = self.config.sample_size == 128
+            # TODO(Sayak, PVP) clean this, for now we use sample size to determine whether to use
+            # additional conditions until we find better name
+            self.adaln_single = AdaLayerNormSingle(inner_dim, use_additional_conditions=self.use_additional_conditions)
+
+        self.caption_projection = None
+        if caption_channels is not None:
+            self.caption_projection = PixArtAlphaTextProjection(in_features=caption_channels, hidden_size=inner_dim)
+
+    def _set_gradient_checkpointing(self, module, value=False):
+        if hasattr(module, "gradient_checkpointing"):
+            module.gradient_checkpointing = value
+
+    def _get_transformer_blocks(
+        self,
+        inner_dim,
+        num_attention_heads,
+        attention_head_dim,
+        dropout,
+        cross_attention_dim,
+        activation_fn,
+        num_embeds_ada_norm,
+        attention_bias,
+        only_cross_attention,
+        double_self_attention,
+        upcast_attention,
+        norm_type,
+        norm_elementwise_affine,
+        norm_eps,
+        attention_type,
+        num_layers,
+    ):
+        transformer_blocks = nn.ModuleList(
+            [
+                BasicTransformerBlock(
+                    inner_dim,
+                    num_attention_heads,
+                    attention_head_dim,
+                    dropout=dropout,
+                    cross_attention_dim=cross_attention_dim,
+                    activation_fn=activation_fn,
+                    num_embeds_ada_norm=num_embeds_ada_norm,
+                    attention_bias=attention_bias,
+                    only_cross_attention=only_cross_attention,
+                    double_self_attention=double_self_attention,
+                    upcast_attention=upcast_attention,
+                    norm_type=norm_type,
+                    norm_elementwise_affine=norm_elementwise_affine,
+                    norm_eps=norm_eps,
+                    attention_type=attention_type,
+                )
+                for d in range(num_layers)
+            ]
+        )
+        return transformer_blocks
+
+    def forward(
+        self,
+        hidden_states: torch.Tensor,
+        encoder_hidden_states: Optional[torch.Tensor] = None,
+        timestep: Optional[torch.LongTensor] = None,
+        added_cond_kwargs: Dict[str, torch.Tensor] = None,
+        class_labels: Optional[torch.LongTensor] = None,
+        cross_attention_kwargs: Dict[str, Any] = None,
+        attention_mask: Optional[torch.Tensor] = None,
+        encoder_attention_mask: Optional[torch.Tensor] = None,
+        return_dict: bool = True,
+    ):
+        """
+        The [`Transformer2DModel`] forward method.
+
+        Args:
+            hidden_states (`torch.LongTensor` of shape `(batch size, num latent pixels)` if discrete, `torch.FloatTensor` of shape `(batch size, channel, height, width)` if continuous):
+                Input `hidden_states`.
+            encoder_hidden_states ( `torch.FloatTensor` of shape `(batch size, sequence len, embed dims)`, *optional*):
+                Conditional embeddings for cross attention layer. If not given, cross-attention defaults to
+                self-attention.
+            timestep ( `torch.LongTensor`, *optional*):
+                Used to indicate denoising step. Optional timestep to be applied as an embedding in `AdaLayerNorm`.
+            class_labels ( `torch.LongTensor` of shape `(batch size, num classes)`, *optional*):
+                Used to indicate class labels conditioning. Optional class labels to be applied as an embedding in
+                `AdaLayerZeroNorm`.
+            cross_attention_kwargs ( `Dict[str, Any]`, *optional*):
+                A kwargs dictionary that if specified is passed along to the `AttentionProcessor` as defined under
+                `self.processor` in
+                [diffusers.models.attention_processor](https://github.com/huggingface/diffusers/blob/main/src/diffusers/models/attention_processor.py).
+            attention_mask ( `torch.Tensor`, *optional*):
+                An attention mask of shape `(batch, key_tokens)` is applied to `encoder_hidden_states`. If `1` the mask
+                is kept, otherwise if `0` it is discarded. Mask will be converted into a bias, which adds large
+                negative values to the attention scores corresponding to "discard" tokens.
+            encoder_attention_mask ( `torch.Tensor`, *optional*):
+                Cross-attention mask applied to `encoder_hidden_states`. Two formats supported:
+
+                    * Mask `(batch, sequence_length)` True = keep, False = discard.
+                    * Bias `(batch, 1, sequence_length)` 0 = keep, -10000 = discard.
+
+                If `ndim == 2`: will be interpreted as a mask, then converted into a bias consistent with the format
+                above. This bias will be added to the cross-attention scores.
+            return_dict (`bool`, *optional*, defaults to `True`):
+                Whether or not to return a [`~models.unets.unet_2d_condition.UNet2DConditionOutput`] instead of a plain
+                tuple.
+
+        Returns:
+            If `return_dict` is True, an [`~models.transformer_2d.Transformer2DModelOutput`] is returned, otherwise a
+            `tuple` where the first element is the sample tensor.
+        """
+        if cross_attention_kwargs is not None:
+            if cross_attention_kwargs.get("scale", None) is not None:
+                logger.warning("Passing `scale` to `cross_attention_kwargs` is deprecated. `scale` will be ignored.")
+        # ensure attention_mask is a bias, and give it a singleton query_tokens dimension.
+        #   we may have done this conversion already, e.g. if we came here via UNet2DConditionModel#forward.
+        #   we can tell by counting dims; if ndim == 2: it's a mask rather than a bias.
+        # expects mask of shape:
+        #   [batch, key_tokens]
+        # adds singleton query_tokens dimension:
+        #   [batch,                    1, key_tokens]
+        # this helps to broadcast it as a bias over attention scores, which will be in one of the following shapes:
+        #   [batch,  heads, query_tokens, key_tokens] (e.g. torch sdp attn)
+        #   [batch * heads, query_tokens, key_tokens] (e.g. xformers or classic attn)
+        if attention_mask is not None and attention_mask.ndim == 2:
+            # assume that mask is expressed as:
+            #   (1 = keep,      0 = discard)
+            # convert mask into a bias that can be added to attention scores:
+            #       (keep = +0,     discard = -10000.0)
+            attention_mask = (1 - attention_mask.to(hidden_states.dtype)) * -10000.0
+            attention_mask = attention_mask.unsqueeze(1)
+
+        # convert encoder_attention_mask to a bias the same way we do for attention_mask
+        if encoder_attention_mask is not None and encoder_attention_mask.ndim == 2:
+            encoder_attention_mask = (1 - encoder_attention_mask.to(hidden_states.dtype)) * -10000.0
+            encoder_attention_mask = encoder_attention_mask.unsqueeze(1)
+
+        # 1. Input
+        if self.is_input_continuous:
+            batch, _, height, width = hidden_states.shape
+            residual = hidden_states
+
+            hidden_states = self.norm(hidden_states)
+            if not self.use_linear_projection:
+                hidden_states = self.proj_in(hidden_states)
+                inner_dim = hidden_states.shape[1]
+                hidden_states = hidden_states.permute(0, 2, 3, 1).reshape(batch, height * width, inner_dim)
+            else:
+                inner_dim = hidden_states.shape[1]
+                hidden_states = hidden_states.permute(0, 2, 3, 1).reshape(batch, height * width, inner_dim)
+                hidden_states = self.proj_in(hidden_states)
+
+        elif self.is_input_vectorized:
+            hidden_states = self.latent_image_embedding(hidden_states)
+        elif self.is_input_patches:
+            height, width = hidden_states.shape[-2] // self.patch_size, hidden_states.shape[-1] // self.patch_size
+            hidden_states = self.pos_embed(hidden_states)
+
+            if self.adaln_single is not None:
+                if self.use_additional_conditions and added_cond_kwargs is None:
+                    raise ValueError(
+                        "`added_cond_kwargs` cannot be None when using additional conditions for `adaln_single`."
+                    )
+                batch_size = hidden_states.shape[0]
+                timestep, embedded_timestep = self.adaln_single(
+                    timestep, added_cond_kwargs, batch_size=batch_size, hidden_dtype=hidden_states.dtype
+                )
+
+        # 2. Blocks
+        if self.caption_projection is not None:
+            batch_size = hidden_states.shape[0]
+            encoder_hidden_states = self.caption_projection(encoder_hidden_states)
+            encoder_hidden_states = encoder_hidden_states.view(batch_size, -1, hidden_states.shape[-1])
+
+        for block in self.transformer_blocks:
+            if self.training and self.gradient_checkpointing:
+
+                def create_custom_forward(module, return_dict=None):
+                    def custom_forward(*inputs):
+                        if return_dict is not None:
+                            return module(*inputs, return_dict=return_dict)
+                        else:
+                            return module(*inputs)
+
+                    return custom_forward
+
+                ckpt_kwargs: Dict[str, Any] = {"use_reentrant": False} if is_torch_version(">=", "1.11.0") else {}
+                hidden_states = torch.utils.checkpoint.checkpoint(
+                    create_custom_forward(block),
+                    hidden_states,
+                    attention_mask,
+                    encoder_hidden_states,
+                    encoder_attention_mask,
+                    timestep,
+                    cross_attention_kwargs,
+                    class_labels,
+                    **ckpt_kwargs,
+                )
+            else:
+                hidden_states = block(
+                    hidden_states,
+                    attention_mask=attention_mask,
+                    encoder_hidden_states=encoder_hidden_states,
+                    encoder_attention_mask=encoder_attention_mask,
+                    timestep=timestep,
+                    cross_attention_kwargs=cross_attention_kwargs,
+                    class_labels=class_labels,
+                )
+
+        # 3. Output
+        if self.is_input_continuous:
+            if not self.use_linear_projection:
+                hidden_states = hidden_states.reshape(batch, height, width, inner_dim).permute(0, 3, 1, 2).contiguous()
+                hidden_states = self.proj_out(hidden_states)
+            else:
+                hidden_states = self.proj_out(hidden_states)
+                hidden_states = hidden_states.reshape(batch, height, width, inner_dim).permute(0, 3, 1, 2).contiguous()
+
+            output = hidden_states + residual
+        elif self.is_input_vectorized:
+            hidden_states = self.norm_out(hidden_states)
+            logits = self.out(hidden_states)
+            # (batch, self.num_vector_embeds - 1, self.num_latent_pixels)
+            logits = logits.permute(0, 2, 1)
+
+            # log(p(x_0))
+            output = F.log_softmax(logits.double(), dim=1).float()
+
+        if self.is_input_patches:
+            if self.config.norm_type != "ada_norm_single":
+                conditioning = self.transformer_blocks[0].norm1.emb(
+                    timestep, class_labels, hidden_dtype=hidden_states.dtype
+                )
+                shift, scale = self.proj_out_1(F.silu(conditioning)).chunk(2, dim=1)
+                hidden_states = self.norm_out(hidden_states) * (1 + scale[:, None]) + shift[:, None]
+                hidden_states = self.proj_out_2(hidden_states)
+            elif self.config.norm_type == "ada_norm_single":
+                shift, scale = (self.scale_shift_table[None] + embedded_timestep[:, None]).chunk(2, dim=1)
+                hidden_states = self.norm_out(hidden_states)
+                # Modulation
+                hidden_states = hidden_states * (1 + scale) + shift
+                hidden_states = self.proj_out(hidden_states)
+                hidden_states = hidden_states.squeeze(1)
+
+            # unpatchify
+            if self.adaln_single is None:
+                height = width = int(hidden_states.shape[1] ** 0.5)
+            hidden_states = hidden_states.reshape(
+                shape=(-1, height, width, self.patch_size, self.patch_size, self.out_channels)
             )
-            for d in range(num_layers)
-        ]
-    )
-    return transformer_blocks
+            hidden_states = torch.einsum("nhwpqc->nchpwq", hidden_states)
+            output = hidden_states.reshape(
+                shape=(-1, self.out_channels, height * self.patch_size, width * self.patch_size)
+            )
+
+        if not return_dict:
+            return (output,)
+
+        return Transformer2DModelOutput(sample=output)
